requirements.txt
librosa
numpy 
keras.models 
flask 
flask_restful
librosa
pandas 
sklearn
joblib
matplotlib
tensorflow